Rough draft of autoconf setup.
Known / potential issues:

* In spite of having a --with-local-cimg option, ./configure might fail
  anyway if CImg.h not found in a system directory.
* Did not test with libpaillier installed system-wide.
* Did a sloppy job with CFLAGS/CXXFLAGS (there are still some things
  hard-coded into the makefile template that probably shouldn't be).
