+++ demo/.gitignore
@@ -1,0 +1,4 @@
+Makefile
+autom4te.cache
+config.status
+configure
