+++ demo/Makefile.in
@@ -1,0 +1,17 @@
+# Top-level Makefile to run other Makefiles...
+
+.PHONY : all
+all :
+	make -C src/
+
+.PHONY : clean
+clean :
+	make -C src/   clean
+
+.PHONY : distclean
+distclean : clean
+	rm -rf autom4te.cache/ config.log config.status configure
+	rm -f @CONFIG_GEN_FILES@
+
+# vim:ft=make:foldmethod=marker:foldmarker={{{,}}}
+
