+++ demo/src/.gitignore
@@ -1,3 +1,4 @@
+Makefile
 demo
 scheme2
