+++ demo/src/client.h
@@ -1,0 +1,4 @@
+#include <vector>
+using std::vector;
+
+vector<int> preproc_radial_sums(char* filename);
